Former-commit-id: 0f31314de7f93f4024103af23aa14580e798ffc0

+++ firebird/chip.py
@@ -24,7 +24,7 @@ def difference(point, interval):
                      on the real number line
     :returns: difference between a point and prior point on an interval.
     """
-   return point % interval
+    return point % interval
 
 
 def near(point, interval, offset):
@@ -97,12 +97,12 @@ def snap(x, y, chip_spec):
     :param chip_spec: parameters for a chip's grid system
     :returns: tuple of chip x & y
   """
-  chip_x  = chip_spec['chip_x']
-  chip_y  = chip_spec['chip_y']
-  shift_x = chip_spec['shift_x']
-  shift_y = chip_spec['shift_y']
-  chip    = point_to_chip(x, y, chip_x, chip_y, shift_x, shift_y)
-  return (long(chip[0]), long(chip[1]))
+    chip_x  = chip_spec['chip_x']
+    chip_y  = chip_spec['chip_y']
+    shift_x = chip_spec['shift_x']
+    shift_y = chip_spec['shift_y']
+    chip    = point_to_chip(x, y, chip_x, chip_y, shift_x, shift_y)
+    return int(chip[0]), int(chip[1])
 
 
 def ids(ulx, uly, lrx, lry, chip_spec):
